requirements.txt
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
fastapi==0.61.0
# Editable Git install with no remote (fastapi-bearer-auth==0.1.0)
-e /Users/yc/gits/fastapi-bearer-auth
flake8==3.8.3
h11==0.9.0
httptools==0.1.1
mccabe==0.6.1
passlib==1.7.2
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
pyflakes==2.2.0
PyJWT==1.7.1
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
